src/omniforge/skills/summarizing-google-docs/requirements.txt
```
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
requests==2.31.0
nltk==3.8.1
```